requirements.txt
requests==2.4.3
Flask==0.9
Flask-SSLify==0.1.5
Flask-OAuthlib==0.9.4
Flask-Cors==3.0.3
oauth2client==4.0.0
python-jose==1.4.0
google-api-python-client==1.6.2
six==1.11.0
pymongo==3.2.2